Fix rgb2gray making errors in metrics computation

+++ src/utils.py
@@ -65,7 +65,7 @@ def rgb2gray(rgb, use_opencv=False):
     else:
         return np.dot(
             rgb[..., :3], [0.299, 0.587, 0.114]
-        ).astype(np.uint8)  # Chosen to comply with OpenCV, see https://docs.opencv.org/3.4/de/d25/imgproc_color_conversions.html
+        ).astype(int)  # Chosen to comply with OpenCV, see https://docs.opencv.org/3.4/de/d25/imgproc_color_conversions.html
 
 def zoom_on_detail(img, top_left=[150,150], width=200, height=200):
     #Zooming on details
